Removed trash + fixed watch sync bug
