+++ AnodaTurnTimerWatch Extension/UserStory/Main/MainController.swift
@@ -8,24 +8,33 @@
 
 import Foundation
 import WatchKit
+import WatchConnectivity
 
-class MainNewController: WKInterfaceController {
+class MainController: WKInterfaceController {
     
     @IBOutlet var playButton: WKInterfaceButton!
     
     @IBOutlet var settingsButton: WKInterfaceButton!
     
+    let session = WCSession.default
+    
     override func awake(withContext context: Any?) {
         super.awake(withContext: context)
+        setupSession()
     }
     
+    private func setupSession() {
+        processAppContext()
+        session.delegate = self
+        session.activate()
+    }
     
     @IBAction private func startTimer() {
-        pushController(withName: Constants.timerControllerClassName, context: nil)
+        pushController(withName: Constants.timerControllerClassName, context: "stub")
     }
     
     @IBAction private func openSettings() {
-        presentController(withNames: [Constants.roundDurationControllerClassName, Constants.beepIntervalController], contexts: nil)
+        presentController(withNames: [Constants.roundDurationControllerClassName, Constants.beepIntervalController], contexts: ["stub", "stub"])
     }
     
 }
